--- src/App.js
@@ -1,12 +1,0 @@
-import React from 'react';
-import HeaderComp from './HeaderComp';
-
-function App() {
-  return (
-    <div>
-        <HeaderComp/>
-    </div>
-  );
-}
-
-export default App;
