+++ src/components/App/App.js
@@ -1,0 +1,8 @@
+import React from 'react';
+import Header from '../Header/Header';
+
+function App() {
+  return <Header/>;
+}
+
+export default App;
