+++ src/css/main.css
@@ -1720,15 +1720,15 @@ button::-moz-focus-inner, input::-moz-focus-inner {
  * # pageheader
  *
  * ------------------------------------------------------------------- */
-.sec-pageheader {
+.s-pageheader {
    min-height: 174px;
-   background-image: url(../images/header-bg.jpg);
+   background-image: url(../images/hero-bg.jpg);
    background-repeat: no-repeat;
    background-position: center center;
    position: relative;
 }
 
-.sec-pageheader::before {
+.s-pageheader::before {
    display: block;
    content: "";
    position: absolute;
@@ -1740,7 +1740,7 @@ button::-moz-focus-inner, input::-moz-focus-inner {
    opacity: .95;
 }
 
-.sec-pageheader-home {
+.s-pageheader--home {
    min-height: 480px;
    padding-top: 222px;
 }
@@ -1752,7 +1752,7 @@ button::-moz-focus-inner, input::-moz-focus-inner {
    width: 100%;
 }
 
-.header_content {
+.header__content {
    position: relative;
    padding: 0 20px;
 }
@@ -1767,14 +1767,14 @@ button::-moz-focus-inner, input::-moz-focus-inner {
 /* -------------------------------------------------------------------
  * ## header logo
  * ------------------------------------------------------------------- */
-.header_logo {
+.header__logo {
    display: inline-block;
    margin: 0;
    padding: 0;
    z-index: 501;
 }
 
-.header_logo a {
+.header__logo a {
    display: block;
    margin: 0 0 0 6px;
    padding: 0;
@@ -1784,7 +1784,7 @@ button::-moz-focus-inner, input::-moz-focus-inner {
    transition: all 0.3s ease-in-out;
 }
 
-.header_logo img {
+.header__logo img {
    width: 265px;
    height: 48px;
 }
@@ -1793,7 +1793,7 @@ button::-moz-focus-inner, input::-moz-focus-inner {
 /* -------------------------------------------------------------------
  * ## header social
  * ------------------------------------------------------------------- */
-.header_social {
+.header__social {
    list-style: none;
    font-size: 23px;
    line-height: 1;
@@ -1804,22 +1804,22 @@ button::-moz-focus-inner, input::-moz-focus-inner {
    top: 12px;
 }
 
-.header_social a {
+.header__social a {
    color: #FFFFFF;
 }
 
-.header_social li {
+.header__social li {
    display: inline-block;
    margin-right: 5px;
    margin-left: 5px;
    padding-left: 0;
 }
 
-.header_social li:last-child {
+.header__social li:last-child {
    margin-right: 0;
 }
 
-.header_social li:first-child {
+.header__social li:first-child {
    margin-left: 0;
 }
 
@@ -1827,7 +1827,7 @@ button::-moz-focus-inner, input::-moz-focus-inner {
 /* -------------------------------------------------------------------
  * ## header-toggle
  * ------------------------------------------------------------------- */
-.header_toggle-menu {
+.header__toggle-menu {
    display: none;
    width: 40px;
    height: 40px;
@@ -1836,7 +1836,7 @@ button::-moz-focus-inner, input::-moz-focus-inner {
    top: 6px;
 }
 
-.header_toggle-menu span {
+.header__toggle-menu span {
    display: block;
    width: 22px;
    height: 2px;
@@ -1854,8 +1854,8 @@ button::-moz-focus-inner, input::-moz-focus-inner {
    color: transparent;
 }
 
-.header_toggle-menu span::before,
-.header_toggle-menu span::after {
+.header__toggle-menu span::before,
+.header__toggle-menu span::after {
    content: '';
    width: 100%;
    height: 100%;
@@ -1864,11 +1864,11 @@ button::-moz-focus-inner, input::-moz-focus-inner {
    left: 0;
 }
 
-.header_toggle-menu span::before {
+.header__toggle-menu span::before {
    top: -8px;
 }
 
-.header_toggle-menu span::after {
+.header__toggle-menu span::after {
    bottom: -8px;
 }
 
@@ -1876,7 +1876,7 @@ button::-moz-focus-inner, input::-moz-focus-inner {
 /* -------------------------------------------------------------------
  * ## header-search
  * ------------------------------------------------------------------- */
-.header_search {
+.header__search {
    display: block;
    text-align: center;
    background: #151515;
@@ -1982,7 +1982,7 @@ body.search-is-visible {
 /* -------------------------------------------------------------------
  * ## header search trigger
  * ------------------------------------------------------------------- */
-.header_search-trigger {
+.header__search-trigger {
    height: 21px;
    min-width: 21px;
    position: absolute;
@@ -1994,7 +1994,7 @@ body.search-is-visible {
    background-size: contain;
 }
 
-.header_search-trigger::before {
+.header__search-trigger::before {
    font-family: "metropolis-semibold", sans-serif;
    font-size: 12px;
    text-transform: uppercase;
